+++ packages/leancode_cubit_utils_cqrs/lib/leancode_cubit_utils_cqrs.dart
@@ -5,5 +5,10 @@ export 'package:leancode_cubit_utils/leancode_cubit_utils.dart';
 
 export 'src/paginated/paginated_query_cubit.dart';
 export 'src/paginated/pre_request/query_pre_request.dart';
-export 'src/request/query_cubit.dart';
-export 'src/request/use_query_cubit.dart';
+export 'src/request/args_query_cubit/args_query_cubit.dart';
+export 'src/request/args_query_cubit/simple_args_query_cubit.dart';
+export 'src/request/args_query_cubit/use_args_query_cubit.dart';
+export 'src/request/query_cubit/query_cubit.dart';
+export 'src/request/query_cubit/simple_query_cubit.dart';
+export 'src/request/query_cubit/use_query_cubit.dart';
+export 'src/request/query_result_handler.dart';