+++ packages/leancode_cubit_utils_cqrs/lib/src/request/query_cubit/simple_query_cubit.dart
@@ -1,0 +1,25 @@
+import 'package:cqrs/cqrs.dart';
+import 'package:leancode_cubit_utils/leancode_cubit_utils.dart';
+import 'package:leancode_cubit_utils_cqrs/src/request/query_cubit/query_cubit.dart';
+import 'package:leancode_cubit_utils_cqrs/src/request/query_cubit/use_query_cubit.dart';
+
+/// Simplified implementation of [QueryCubit] created in order to be used by
+/// [useQueryCubit].
+class SimpleQueryCubit<TOut> extends QueryCubit<TOut, TOut> {
+  /// Creates a new [SimpleQueryCubit].
+  SimpleQueryCubit(
+    super.loggerTag,
+    this._customRequest, {
+    super.isEmpty,
+    super.requestMode,
+  });
+
+  /// The request to be executed.
+  final Request<QueryResult<TOut>> _customRequest;
+
+  @override
+  Future<QueryResult<TOut>> request() => _customRequest();
+
+  @override
+  TOut map(TOut data) => data;
+}