+++ packages/leancode_cubit_utils_cqrs/lib/src/request/query_result_handler.dart
@@ -3,7 +3,14 @@ import 'package:leancode_cubit_utils/leancode_cubit_utils.dart';
 
 /// A mixin that handles the result of a CQRS query.
 mixin QueryResultHandler<TRes, TOut>
-    on BaseRequestCubit<QueryResult<TRes>, TRes, TOut, QueryError> {
+    on BaseRequestCubit<QueryResult<TRes>, TOut, QueryError> {
+  /// Maps the given [data] to the output type [TOut].
+  TOut map(TRes data);
+
+  /// Override this to implement logic to check if the given [data] is empty.
+  /// When this method returns true, the cubit will emit an [RequestEmptyState]
+  bool isEmpty(TOut data) => false;
+
   @override
   Future<RequestState<TOut, QueryError>> handleResult(
     QueryResult<TRes> result,
@@ -12,7 +19,7 @@ mixin QueryResultHandler<TRes, TOut>
       final mappedData = map(data);
       if (isEmpty(mappedData)) {
         logger.warning('Query success but data is empty');
-        return RequestEmptyState();
+        return RequestEmptyState(mappedData);
       }
       logger.info('Query success. Data: $data');
       return RequestSuccessState(mappedData);
@@ -30,22 +37,4 @@ mixin QueryResultHandler<TRes, TOut>
       return RequestErrorState();
     }
   }
-}
-
-/// Base class for all request cubits that perform a CQRS query
-/// and do not require any arguments.
-abstract class QueryCubit<TRes, TOut>
-    extends RequestCubit<QueryResult<TRes>, TRes, TOut, QueryError>
-    with QueryResultHandler {
-  /// Creates a new [RequestCubit] with the given [requestMode].
-  QueryCubit(super.loggerTag, {super.requestMode});
-}
-
-/// Base class for all request cubits that perform a CQRS query
-/// and require arguments.
-abstract class ArgsQueryCubit<TArgs, TRes, TOut>
-    extends ArgsRequestCubit<TArgs, QueryResult<TRes>, TRes, TOut, QueryError>
-    with QueryResultHandler {
-  /// Creates a new [ArgsRequestCubit] with the given [requestMode].
-  ArgsQueryCubit(super.loggerTag, {super.requestMode});
 }