+++ packages/leancode_cubit_utils_cqrs/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.5.0
+
+* **BREAKING CHANGE:** Remove `useQueryWithEmptyCubit` and `SimpleQueryWithEmptyCubit`.
+* **BREAKING CHANGE:** Remove `useArgsQueryWithEmptyCubit` and `SimpleArgsQueryWithEmptyCubit`.
+* Start using `isEmpty` checker of `useQueryCubit` and pass it `SimpleQueryCubit`.
+* Add `isEmpty` checker to `useArgsQueryCubit` and `SimpleArgsQueryCubit`.
+
 ## 0.4.0
 
 * Update `leancode_cubit_utils` dependency to `^0.4.0` (includes breaking change: `builder` parameter renamed to `onSuccess` in `RequestCubitBuilder`)